requirements.txt
# Todo应用依赖包
requests>=2.28.0
python-dotenv>=0.19.0
pyinstaller>=5.0.0

